requirements.txt
flet>=0.8.0
transformers
torch
chromadb
spacy
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
bitsandbytes
accelerate
qwen-vl-utils
pillow
einops
timm
pathlib
python-dotenv